+++ workspace/src/barc/src/barc_simulator.jl
@@ -27,15 +27,15 @@ using JLD
 
 include("barc_lib/simModel.jl")
 
-u_current = zeros(Float64,2,1)      # msg ECU is Float32 !
+u_current = zeros(Float32,2,1)      # msg ECU is Float32 !
 
 t = 0
 
 
 # This type contains measurement data (time, values and a counter)
 type Measurements{T}
-    i::Int64          # measurement counter
-    t::Array{Float64}       # time data
+    i::Int32          # measurement counter
+    t::Array{Float32}       # time data
     z::Array{T}       # measurement values
 end
 # This function cleans the zeros from the type above once the simulation is finished
@@ -45,12 +45,12 @@ function clean_up(m::Measurements)
 end
 
 buffersize = 60000
-gps_meas = Measurements{Float64}(0,zeros(buffersize),zeros(buffersize,2))
-imu_meas = Measurements{Float64}(0,zeros(buffersize),zeros(buffersize,2))
+gps_meas = Measurements{Float32}(0,zeros(buffersize),zeros(buffersize,2))
+imu_meas = Measurements{Float32}(0,zeros(buffersize),zeros(buffersize,2))
 est_meas = Measurements{Float32}(0,zeros(buffersize),zeros(Float32,buffersize,4))
 est_meas_dyn = Measurements{Float32}(0,zeros(buffersize),zeros(Float32,buffersize,6))
-cmd_log  = Measurements{Float64}(0,zeros(buffersize),zeros(buffersize,2))
-z_real   = Measurements{Float64}(0,zeros(buffersize),zeros(buffersize,4))
+cmd_log  = Measurements{Float32}(0,zeros(buffersize),zeros(buffersize,2))
+z_real   = Measurements{Float32}(0,zeros(buffersize),zeros(buffersize,4))
 
 z_real.t[1]   = time()
 imu_meas.t[1] = time()
